# TravelGuideApi/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Travel Guide API",
    "description": "API for the Travel Guide mobile application",
    "contact": {
      "name": "Travel Guide Team"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Reviews": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "PageNumber",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "PageSize",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "SortBy",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "SortOrder",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ReviewDtoPaginatedResponseApiResponse"
            }
          }
        }
      }
    },
    "/api/Reviews/{id}": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ReviewDtoApiResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Reviews"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ReviewUpdateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ReviewDtoApiResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Reviews"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/BooleanApiResponse"
            }
          }
        }
      }
    },
    "/api/Destinations/{destinationId}/Reviews": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "destinationId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "PageNumber",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "PageSize",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "SortBy",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "SortOrder",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ReviewDtoPaginatedResponseApiResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Reviews"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "destinationId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ReviewCreateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ReviewDtoApiResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "BooleanApiResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "type": "boolean"
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "additionalProperties": false
    },
...
    "ReviewCreateDto": {
      "type": "object",
      "properties": {
        "destinationId": {
          "format": "int32",
          "type": "integer"
        },
        "reviewerName": {
          "type": "string"
        },
        "reviewText": {
          "type": "string"
        },
        "rating": {
          "format": "float",
          "type": "number"
        }
      },
      "additionalProperties": false
    },
    "ReviewDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "destinationId": {
          "format": "int32",
          "type": "integer"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "reviewerName": {
          "type": "string"
        },
        "reviewText": {
          "type": "string"
        },
        "rating": {
          "format": "float",
          "type": "number"
        },
        "reviewDate": {
          "format": "date-time",
          "type": "string"
        },
        "destinationName": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "ReviewDtoApiResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/ReviewDto"
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "additionalProperties": false
    },
    "ReviewDtoPaginatedResponse": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReviewDto"
          }
        },
        "totalCount": {
          "format": "int32",
          "type": "integer"
        },
        "pageNumber": {
          "format": "int32",
          "type": "integer"
        },
        "pageSize": {
          "format": "int32",
          "type": "integer"
        },
        "totalPages": {
          "format": "int32",
          "type": "integer"
        },
        "hasPrevious": {
          "type": "boolean"
        },
        "hasNext": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
    "ReviewDtoPaginatedResponseApiResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/ReviewDtoPaginatedResponse"
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "additionalProperties": false
    },
    "ReviewUpdateDto": {
      "type": "object",
      "properties": {
        "reviewText": {
          "type": "string"
        },
        "rating": {
          "format": "float",
          "type": "number"
        }
      },
      "additionalProperties": false
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}